refactor(chat): split hub event forwarder helpers
